+++ components/layouts/layout.jsx
@@ -1,5 +1,5 @@
 import Head from 'next/head'
-import styles from '../styles/layout.module.css'
+import styles from '../../styles/layout.module.css'
 
 export default function Layout({ children }) {
     return(
@@ -8,8 +8,11 @@ export default function Layout({ children }) {
                 <meta charSet="UTF-8" />
                 <link rel="icon" type="image/favicon" href="/favicon.ico" />
                 <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-                <title>Discord Timestamp Generator</title>
-                <meta name="description" content="Create your Discord timestamps here! Discord in Time generates beatiful dynamic date-time display in your Discord messages. Choose your time and copy the code below."/>
+                <title>Discord Timestamp Generator | Discord in Time</title>
+                <meta 
+                    name="description" 
+                    content="Create your Discord timestamps here! Discord in Time generates beatiful dynamic date-time display in your Discord messages. Choose your time and copy the code below." 
+                    key="desc"/>
                 <meta name="keywords" content="discord time, discord time messages, discord timestamp, discord, timestamp, generator"/>
                 <meta property='og:title' content='Discord Timestamp Generator'/>
                 <meta property="og:type" content="website" />
